requirements.txt
opencv-python==4.8.1.78
openai==1.3.7
python-dotenv==1.0.0
